Struct r3_kernel::TaskCb

#[repr(C)]
pub struct TaskCb<Traits: PortThreading, PortTaskState: 'static = <Traits as PortThreading>::PortTaskState, TaskPriority: 'static = <Traits as KernelCfg1>::TaskPriority, TaskReadyQueueData: 'static = <<Traits as KernelCfg1>::TaskReadyQueue as Queue<Traits>>::PerTaskData> {
    pub port_task_state: PortTaskState,
    pub attr: &'static TaskAttr<Traits, TaskPriority>,
    /* private fields */
}

Task control block - the state data of a task.

Fields
port_task_state: PortTaskState

Get a reference to PortTaskState in the task control block.

This is guaranteed to be placed at the beginning of the struct so that assembly code can refer to this easily.

attr: &'static TaskAttr<Traits, TaskPriority>

The static properties of the task.