Module r3_kernel::utils

Utility

This module is exempt from the API stability guarantee unless specified otherwise. It’s exposed mostly because it’s needed by macros.

Modules

binary_heap

Binary heap with a contextful comparator and position tracking

mem

Structs

AlignedStorage

Untyped storage of the specified size and alignment. This is analogous to C++’s std::aligned_storage_t.

Frozen

Erases interior mutability by preventing reference forming.

RawCell

Like UnsafeCell, but implements Sync.

Constants

FIXED_PRIO_BITMAP_MAX_LEN

The maximum bit count supported by FixedPrioBitmap.

Traits

BinInteger

Integral types with efficient binary operations.

BinUInteger

Unsigned integral types with efficient binary operations.

Init

Trait for types having a constant default value. This is essentially a constant version of Default.

PrioBitmap

Trait for FixedPrioBitmap.

Zeroable

Trait for types that can be safely created with zeroed.

Type Definitions

FixedPrioBitmap

A bit array structure supporting logarithmic-time bit scan operations.

UIntegerWithBound

Get the smallest unsigned integer type capable of representing the specified value.

Derive Macros

Zeroable

Derive the Zeroable trait for a struct